refactor: change directories tree
