+++ chat/index.html
@@ -100,6 +100,6 @@
         </div>
       </div>
     </div>
-    <script type="module" src="/src/main.js"></script>
+    <script type="module" src="/src/chat/main.js"></script>
   </body>
 </html>
