+++ translate/index.html
@@ -84,6 +84,6 @@
         </div>
       </div>
     </div>
-    <script type="module" src="/src/translator.js"></script>
+    <script type="module" src="/src/translator/main.js"></script>
   </body>
 </html>
